==> 2_presence_archetype.Rmd <==
---
title: "Description de la présence de chaque Archetype"
output: 
  html_document:
date: "`r Sys.Date()`"
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "outpout/result") })
---

```{r setup, include=FALSE, message = FALSE, warning = FALSE} 
knitr::opts_chunk$set(
  echo = FALSE,
  message = FALSE,
  warning = FALSE
)
#

library("rjson")
library(lubridate)
library(kableExtra)
library(tidyverse)
library(plotly)
# devtools::install_github("hoesler/rwantshue")
library("rwantshue")
library(htmltools)

# css: TOC.css
source("S2_Source_mtg_new_card.R",local = TRUE)

loaded_base_data <- readRDS("data/data_meta_en_cours.rds")
# Réfléchir pour les couleurs et meilleurs distances
# Manque ajouter du texte après les % pour la variation
conflicted::conflicts_prefer(dplyr::filter)
```


```{r}
knitr::read_chunk("S2_Source_mtg_new_card.R")
```


```{r, Introduction_chunk_2_Deck_analysis, results='asis'}

```

```{r}
# Generate different color scheme for close archetype
scheme <- iwanthue(seed = 42, force_init = TRUE) # recreate with a seed

Presence_df_base_no_filter <- loaded_base_data %>%
  filter(Tournament != "Modern League") %>%
  group_by(Archetype) %>%
  mutate(
    Archetype_count = n()
  ) %>%
  arrange(Archetype_count)


filter_archetype_count <- min(50, 
 sum(unique(Presence_df_base_no_filter$Archetype_count))/100 )

Presence_df_base <- Presence_df_base_no_filter %>%
  filter(Archetype_count > filter_archetype_count) %>%
  ungroup()



Presence_df_base_all_data <- Presence_df_base %>%
  mutate(
    Archetype =
      factor(Archetype,
        level = unique(.$Archetype)
      ),
    Rank = as.numeric(
      factor(Archetype,
        level = rev(unique(.$Archetype))
      )
    ),
    Base_Archetype =

      factor(Base_Archetype,
        level = unique(.$Base_Archetype)
      ),

    # A reflechir ou tu mets ça selon les filters
    Archetype_percent = Archetype_count / nrow(.)
  )

# Plot color scales
colors_scales_presence <- scheme$hex(
  length(levels(Presence_df_base_all_data$Base_Archetype))
)


one_week_inline <- Presence_df_base_all_data %>%
  filter(
    Week > (max(Week) - 2)
  )

two_week_inline <- Presence_df_base_all_data %>%
  filter(
    Week > (max(Week) - 3)
  )


one_month_inline <- Presence_df_base_all_data %>%
  filter(
    Week > (max(Week) - 5)
  )
```



```{r}
# a <- DF_presence_fun(
#       df_base = Presence_df_base,
#   time_limit = Inf,
#   compare_time_limit = NULL
#   )

```

```{r}
spaghetti_plot_df <- Presence_df_base %>%
  arrange(desc(Archetype_count)) %>%
  mutate(
    Archetype =
      factor(Archetype,
        level = unique(.$Archetype)
      ),
    Base_Archetype =

      factor(str_remove(Base_Archetype, " _fallback"),
        level = unique(str_remove(.$Base_Archetype, " _fallback"))
      )
  ) %>%
  select(Week, Meta, Archetype, Base_Archetype) %>%
  # add_count(Week,Archetype,
  #           name = "Count_arch") %>%
  group_by(Week, Archetype) %>%
  mutate(Count_arch = n()) %>%
  ungroup() %>%
  group_by(Week) %>%
  mutate(Week_deck_number = n()) %>%
  ungroup() %>%
  mutate(Archetype_percent = (Count_arch / Week_deck_number)) %>%
  # add_count(Week,Base_Archetype,
  #           name = "Count_base_arch") %>%
  group_by(Week, Base_Archetype) %>%
  mutate(Count_base_arch = n()) %>%
  ungroup() %>%
  mutate(Base_Archetype_percent = (Count_base_arch / Week_deck_number)) %>%
  distinct()



Low_Arch_spaghetti <- as.character(
  spaghetti_plot_df %>%
  distinct(Week, Archetype, Archetype_percent) %>%
  group_by(Archetype) %>%
  # old based on mean presence
  # summarise(Archetype_percent = sum(Archetype_percent) / max(Week), ) %>%
  # now based on max presence
  summarise(Archetype_percent = max(Archetype_percent)) %>%
   filter(Archetype_percent < 0.025) %>% pull(Archetype)
)


Low_base_Arch_spaghetti <- as.character(
  spaghetti_plot_df %>%
  distinct(Week, Base_Archetype, Base_Archetype_percent) %>%
  group_by(Base_Archetype) %>%
  # old based on mean presence
  # summarise(Base_Archetype_percent = sum(Base_Archetype_percent) / max(Week), ) %>%
  # now based on max presence
  summarise(Base_Archetype_percent = max(Base_Archetype_percent)) %>%
  
  filter(Base_Archetype_percent < 0.025) %>% pull(Base_Archetype)
  )
```



# Presence over Time 

## Archetype 
```{r out.width="100%"}
plot_archetype_spaghetti <- (ggplot(
  spaghetti_plot_df,
  aes(
    x = Week,
    y = Archetype_percent,
    color = Archetype,
    text = paste(
      "Archetype: ", Archetype, "<br>", # Archetype name
      "Archetype percent: ", round(Archetype_percent * 100, 2), " %", "<br>",
      sep = ""
    ),
    group = 1
  )
) +
  geom_line() +
  geom_point() +
  scale_x_continuous("Week",
    breaks = unique(spaghetti_plot_df$Week)
  ) +
  scale_color_manual(
    values = colors_scales_presence[1:length(
      unique(spaghetti_plot_df$Base_Archetype)
    )]
  ) +
  ylab("Archetype presence") +
  scale_y_continuous(labels = function(x) scales::percent(x))
) %>%
  ggplotly(tooltip = c("text"), height = (480 * 2.25), width = (640 * 2.25))


plot_archetype_spaghetti$x$data <- lapply(
  plot_archetype_spaghetti$x$data, function(y) {
    if (y$name %in% Low_Arch_spaghetti) {
      y$visible <- "legendonly"
    }


    return(y)
  }
)


plot_archetype_spaghetti
```


## Base Archetype

```{r out.width="100%"}
plot_base_archetype_spaghetti <- (ggplot(
  spaghetti_plot_df,
  aes(
    x = Week,
    y = Base_Archetype_percent,
    color = Base_Archetype,
    text = paste(
      "Archetype: ", Base_Archetype, "<br>", # Archetype name
      "Archetype percent: ", round(Base_Archetype_percent * 100, 2), " %", "<br>",
      sep = ""
    ),
    group = 1
  )
) +
  geom_line() +
  geom_point() +
  scale_x_continuous("Week",
    breaks = unique(spaghetti_plot_df$Week)
  ) +
  scale_color_manual(
    values = colors_scales_presence[1:length(
      unique(spaghetti_plot_df$Base_Archetype)
    )]
  ) +
  ylab("Archetype presence") +
  scale_y_continuous(labels = function(x) scales::percent(x))
) %>%
  ggplotly(tooltip = c("text"), height = (480 * 2.25), width = (640 * 2.25))



plot_base_archetype_spaghetti$x$data <- lapply(
  plot_base_archetype_spaghetti$x$data, function(y) {
    if (y$name %in% Low_base_Arch_spaghetti) {
      y$visible <- "legendonly"
    }


    return(y)
  }
)
plot_base_archetype_spaghetti
```




# All Data 

* All data : `r format(as.Date(min(Presence_df_base$Date)), "%d-%m-%Y")` to `r format(as.Date(max(Presence_df_base$Date)), format = "%d/%m/%Y")`. 
* Duration `r round(max(difftime(Sys.Date(),Presence_df_base$Date,units = "weeks")),0)` weeks.
* Include meta : `r unique(Presence_df_base_all_data$Meta)`.




```{r out.width="100%"}
plot_presence_fun(
  df_base = Presence_df_base,
  color_scheme = colors_scales_presence,
  time_limit = Inf,
  compare_time_limit = NULL
)
```

# 1 Month

* 1 Month data : `r format(as.Date(min(one_month_inline$Date)), "%d-%m-%Y")` to `r format(as.Date(max(one_month_inline$Date)), "%d-%m-%Y")`.
* Duration `r round(max(difftime(Sys.Date(),one_month_inline$Date,units = "days")),0)` days.
* Include meta : `r unique(one_month_inline$Meta)`.


```{r out.width="100%"}
plot_presence_fun(
  df_base = Presence_df_base,
  color_scheme = colors_scales_presence,
  time_limit = 5,
  compare_time_limit = Inf
)
```


# 2 Weeks 

* 2 Weeks data : `r format(as.Date(min(two_week_inline$Date)), "%d-%m-%Y")` to `r format(as.Date(max(two_week_inline$Date)), "%d-%m-%Y")`.
* Duration `r round(max(difftime(Sys.Date(),two_week_inline$Date,units = "days")),0)` days.
* Include meta : `r unique(two_week_inline$Meta)`.


```{r ,out.height = "100%",out.width="100%"}
plot_presence_fun(
  df_base = Presence_df_base,
  color_scheme = colors_scales_presence,
  time_limit = 3,
  compare_time_limit = 5
)
```


# 1 Weeks 

* 1 Weeks data : `r format(as.Date(min(one_week_inline$Date)), "%d-%m-%Y")` to `r format(as.Date(max(one_week_inline$Date)), "%d-%m-%Y")`.
* Duration `r round(max(difftime(Sys.Date(),one_week_inline$Date,units = "days")),0)` days.
* Include meta : `r unique(one_week_inline$Meta)`.


```{r out.width="100%"}
plot_presence_fun(
  df_base = Presence_df_base,
  color_scheme = colors_scales_presence,
  time_limit = 2,
  compare_time_limit = 5
)
```







```{r}
a <- DF_presence_fun(
      df_base = Presence_df_base,
  time_limit = Inf,
  compare_time_limit = NULL
  )

```

```{r}
unique(loaded_base_data$Tournament)
a <- loaded_base_data %>% 
  group_by(TournamentFile) %>% 
  # mutate(min_wins = min(Wins,na.rm = TRUE)) %>% 
  # filter(str_detect(Tournament,
  #              "Modern Challenge \\d+|Modern Preliminary|Modern Showcase Challenge|Modern Super Qualifier|Modern Last Chance|Modern Showcase Qualifier"))%>% view()
  # rowwise() 
  mutate( a =  if_else(
    str_detect(Tournament,
               "Modern Challenge \\d+|Modern Preliminary|Modern Showcase Challenge|Modern Super Qualifier|Modern Last Chance|Modern Showcase Qualifier"),
    Wins - min(Wins,na.rm = TRUE),Wins
    
    ),.before = Wins
    
    ) 




a %>% group_by(TournamentFile) %>% 
  summarise(winrate_base = sum(Wins,na.rm = TRUE)/(sum(Losses,na.rm = TRUE)+sum(Wins,na.rm = TRUE)),
            patch_mtg0 = sum(a,na.rm = TRUE)/(sum(Losses,na.rm = TRUE) + sum(a,na.rm = TRUE))
            
            ) %>% 
  mutate(confirm = winrate_base== patch_mtg0) %>% 
  
  view()
```
